supermercado_django/requirements.txt
Django>=4.2
djangorestframework>=3.14
mysqlclient>=2.1
python-dotenv>=1.0
django-cors-headers>=3.15
reportlab>=4.0
